refactor delete pages/marketplace
