+++ src/options/pages/blueprints/utils/useInstall.ts
